requirements.txt
gspread>=6.0.0
google-auth>=2.20.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
